精度调节测试

+++ ttsreal2.py
@@ -174,7 +174,7 @@ class VoitsTTS(BaseTTS):
                     incomplete_chunk = b""
                 if len(chunk) > 0:
                     # 将字节流转换为浮点数流
-                    stream = np.frombuffer(chunk, dtype=np.int16).astype(np.float16) / 32767
+                    stream = np.frombuffer(chunk, dtype=np.int16).astype(np.float32) / 32767
                     stream = resampy.resample(x=stream, sr_orig=32000, sr_new=self.sample_rate)
                     streamlen = stream.shape[0]
                     # print(f"Received audio chunk size: {len(chunk)} stream size: {streamlen}")
